restructure folder setup
