+++ src/llamacpp-adapter.ts
